requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
opencv-contrib-python==4.5.3.56
panda==0.3.1
pandas==1.2.2
Pillow==8.1.0
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
soupsieve==2.2
textblob==0.15.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.56.2
tweepy==3.10.0
urllib3==1.26.3
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
